refactor: separated code into components
